# service-desk-ai-platform-backend/openapi.yaml
openapi: 3.0.3
info:
  title: AI Service Desk Knowledge Intelligence Platform REST API
  description: Enterprise ServiceNow RAG Knowledge Intelligence & Pre-Incident Deflection Engine
  version: 2.5.0-SNAPSHOT
  contact:
    name: Platform Engineering
    email: platform-team@enterprise.com
servers:
  - url: http://localhost:8080
    description: Local Container Runtime
  - url: https://ai-servicedesk.internal.enterprise.com
    description: Production Enterprise Environment
paths:
  /api/v1/suggestions/resolve:
    post:
      tags:
        - AI Suggestion Engine
      summary: Generate AI resolution for pre-ticket query
      operationId: resolveIncident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveIncidentRequest'
      responses:
        '200':
          description: Successful AI resolution generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuggestionResponse'
        '400':
          description: Invalid request parameters
        '500':
          description: Internal server error
...
components:
  schemas:
    ResolveIncidentRequest:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
          example: "GlobalProtect VPN Connection Failed"
        description:
          type: string
          example: "User receives error Server Certificate Invalid when connecting to VPN"
        callerEmail:
          type: string
          example: "user@enterprise.com"
        userDepartment:
          type: string
          example: "Sales"
        category:
          type: string
          example: "Network"
        minConfidenceThreshold:
          type: integer
          example: 75
    SuggestionResponse:
      type: object
      properties:
        suggestionId:
          type: string
          example: "sug-92810"
        queryTitle:
          type: string
        recommendedTitle:
          type: string
        summaryResolution:
          type: string
        stepByStepInstructions:
          type: array
          items:
            type: string
        codeOrCommandSnippet:
          type: string
        confidenceScore:
          type: integer
          example: 92
        confidenceBand:
          type: string
          example: "VERY_HIGH"
        deflectionSuccessful:
          type: boolean
          example: true
        sourcesCount:
          type: integer
          example: 3
        generatedByModel:
          type: string
          example: "gemini-3.6-flash"
...